+++ client/app/pods/paper-tracker/index/route.js
@@ -7,17 +7,25 @@ export default Ember.Route.extend({
     page: {
       refreshModel: true
     },
+    orderBy: {
+      refreshModel: true
+    },
+    orderDir: {
+      refreshModel: true
+    },
+    query: {
+      refreshModel: true
+    },
   },
 
   model(params) {
     return this.get('restless').get('/api/paper_tracker', params).then((data)=> {
+      console.log(data);
       this.prepMetaData(data);
       this.store.pushPayload('paper', data);
       let paperIds = data.papers.mapBy('id');
-      return this.store.all('paper').filter(function(p) {
-        if(paperIds.contains( parseInt(p.get('id')) )) {
-          return p;
-        }
+      return _.collect(paperIds, (id) => {
+        return this.store.find('paper', id);
       });
     });
   },
@@ -42,4 +50,12 @@ export default Ember.Route.extend({
     controller.set('totalCount', this.get('metaData.totalCount'));
     controller.set('perPage', this.get('metaData.perPage'));
   },
+
+  actions: {
+    didTransition: function() {
+      //keeps search box up to date if entering url cold
+      this.controller.set('queryInput', this.controller.get('query'));
+      return true; // Bubble the didTransition event
+    }
+  }
 });